requirements.txt
# Flask Web Framework
Flask==2.3.3
Flask-CORS==4.0.0
Flask-JWT-Extended==4.5.3
Flask-Limiter==3.5.0
Flask-Mail==0.9.1
gunicorn==21.2.0
uvicorn==0.24.0

# Database
SQLAlchemy==2.0.23
psycopg2-binary==2.9.9
alembic==1.13.1
pgvector==0.1.8

# Redis
redis==5.0.1
hiredis==2.2.3

# OpenAI and AI
openai==1.3.5
tiktoken==0.5.1

# Document Processing
langchain==0.0.340
sentence-transformers==2.2.2
PyPDF2==3.0.1
docx2txt==0.8
python-docx==1.1.2
BeautifulSoup4==4.12.2
html2text==2020.1.16
chardet==5.2.0
pdf2image==1.16.3
pytesseract==0.3.10

# Data Processing
numpy==1.24.3
pandas==2.1.3
scipy==1.11.4

# Authentication and Security
bcrypt==4.1.2
PyJWT==2.8.0
cryptography==41.0.7
passlib==1.7.4

# HTTP and API
requests==2.31.0
httpx==0.25.2
aiohttp==3.9.1
fastapi==0.104.1
pydantic==2.5.0

# File and Image Processing
Pillow==10.1.0
openpyxl==3.1.2
xlsxwriter==3.1.9

# Utilities
python-dotenv==1.0.0
click==8.1.7
rich==13.7.0
tqdm==4.66.1
scheduler==0.8.7
apscheduler==3.10.4

# Monitoring and Logging
prometheus-client==0.19.0
structlog==23.2.0
sentry-sdk==1.38.0

# Development and Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
mypy==1.7.1

# Documentation
sphinx==7.2.6
sphinx-rtd-theme==1.3.0
